mod template and mod prime fact

+++ libraries/NTL/Prime_Factorize.cpp
@@ -15,7 +15,7 @@ vector<ll> PrimeFact(ll n){
             continue;
         }
         bool prime_flag = false;
-        for(int i = 2;i*i <= n;i++){
+        for(ll i = 2;i*i <= n;i++){
             if(n%i==0){
                 res.emplace_back(i); n/=i;
                 prime_flag = true;
